results/rebuttal/run-20231111-034650_LLMatic-zero_shot_donnemartin-system-design-oop-bookclub_0_20-runs_log/bookclub/2/LLMatic/app.py
from flask import Flask, request

app = Flask(__name__)

# Mock database
DATABASE = {
	'users': {},
	'book_clubs': {},
	'meetings': {},
	'discussions': {},
	'book_selections': {},
	'user_profiles': {},
	'recommendations': {},
	'admin_actions': {},
	'notifications': {},
	'resources': {}
}

class User:
	def __init__(self, username, password):
		self.username = username
		self.password = password

	def create_user(self):
		DATABASE['users'][self.username] = self.password

	def authenticate_user(self, password):
		return DATABASE['users'].get(self.username) == password

	def get_user_info(self):
		return {'username': self.username, 'password': DATABASE['users'].get(self.username)}

class Profile:
	def __init__(self, username):
		self.username = username
		self.books = []
		self.following = []

	def view_profile(self):
		return DATABASE['user_profiles'].get(self.username)

	def edit_profile(self, new_info):
		DATABASE['user_profiles'][self.username].update(new_info)

	def list_books(self):
		return self.books

	def follow_user(self, user_to_follow):
		self.following.append(user_to_follow)

	def create_profile(self):
		DATABASE['user_profiles'][self.username] = {'books': self.books, 'following': self.following}

class BookClub:
	def __init__(self, name, admin):
		self.name = name
		self.admin = admin
		self.members = [admin]

	def create_book_club(self):
		DATABASE['book_clubs'][self.name] = {'admin': self.admin, 'members': self.members}

	def add_member(self, member):
		if member not in DATABASE['book_clubs'][self.name]['members']:
			DATABASE['book_clubs'][self.name]['members'].append(member)

	def remove_member(self, member):
		if member in DATABASE['book_clubs'][self.name]['members']:
			DATABASE['book_clubs'][self.name]['members'].remove(member)

	def get_book_club_info(self):
		return DATABASE['book_clubs'].get(self.name)

class Meeting:
	def __init__(self, meeting_id, book_club, date, time):
		self.meeting_id = meeting_id
		self.book_club = book_club
		self.date = date
		self.time = time

	def schedule_meeting(self):
		DATABASE['meetings'][self.meeting_id] = {'book_club': self.book_club, 'date': self.date, 'time': self.time}

	def send_reminder(self):
		# In a real application, this would send a reminder to all members of the book club
		return 'Reminder sent'

	def integrate_with_calendar(self):
		# In a real application, this would integrate the meeting with an external calendar app
		return 'Integrated with calendar'

class Discussion:
	def __init__(self, discussion_id, book_club, topic):
		self.discussion_id = discussion_id
		self.book_club = book_club
		self.topic = topic
		self.comments = []
		self.resources = []

	def create_discussion(self):
		DATABASE['discussions'][self.discussion_id] = {'book_club': self.book_club, 'topic': self.topic, 'comments': self.comments, 'resources': self.resources}

	def post_comment(self, comment):
		DATABASE['discussions'][self.discussion_id]['comments'].append(comment)

	def upload_resource(self, resource):
		DATABASE['discussions'][self.discussion_id]['resources'].append(resource)

class Book:
	def __init__(self, book_id, title, author):
		self.book_id = book_id
		self.title = title
		self.author = author
		self.votes = 0

	def suggest_book(self):
		DATABASE['book_selections'][self.book_id] = {'title': self.title, 'author': self.author, 'votes': self.votes}

	def vote_for_book(self):
		DATABASE['book_selections'][self.book_id]['votes'] += 1

	def get_book_info(self):
		return DATABASE['book_selections'].get(self.book_id)

class Recommendation:
	def __init__(self, username):
		self.username = username

	def generate_recommendation(self):
		# In a real application, this would use a recommendation algorithm
		# For simplicity, we'll just recommend the most popular book
		most_popular_book = max(DATABASE['book_selections'].items(), key=lambda x: x[1]['votes'])
		DATABASE['recommendations'][self.username] = most_popular_book[0]
		return most_popular_book[0]

	def highlight_popular_book(self):
		# In a real application, this would use a popularity algorithm
		# For simplicity, we'll just highlight the most voted book
		most_voted_book = max(DATABASE['book_selections'].items(), key=lambda x: x[1]['votes'])
		return most_voted_book[0]

class Admin:
	def __init__(self, username):
		self.username = username

	def manage_book_club(self, book_club_name, action, member=None):
		book_club = BookClub(book_club_name, self.username)
		if action == 'add_member':
			book_club.add_member(member)
		elif action == 'remove_member':
			book_club.remove_member(member)
		elif action == 'delete_book_club':
			DATABASE['book_clubs'].pop(book_club_name, None)
		return 'Action performed successfully'

	def manage_user_account(self, user_to_manage, action):
		if action == 'delete_user':
			DATABASE['users'].pop(user_to_manage, None)
			DATABASE['user_profiles'].pop(user_to_manage, None)
		return 'Action performed successfully'

	def remove_inappropriate_content(self, discussion_id):
		DATABASE['discussions'].pop(discussion_id, None)
		return 'Inappropriate content removed'

	def view_analytics(self):
		# In a real application, this would return detailed analytics
		# For simplicity, we'll just return the number of users and book clubs
		return {'num_users': len(DATABASE['users']), 'num_book_clubs': len(DATABASE['book_clubs'])}

class Notification:
	def __init__(self, username, message):
		self.username = username
		self.message = message

	def setup_notification(self):
		DATABASE['notifications'][self.username] = self.message

	def send_email_alert(self):
		# In a real application, this would send an email
		return 'Email alert sent'

@app.route('/')
def home():
	return 'Welcome to the Book Club App!'

@app.route('/register', methods=['POST'])
def register():
	data = request.get_json()
	user = User(data['username'], data['password'])
	user.create_user()
	profile = Profile(data['username'])
	profile.create_profile()
	return {'message': 'User and profile created successfully'}, 201

@app.route('/login', methods=['POST'])
def login():
	data = request.get_json()
	user = User(data['username'], data['password'])
	if user.authenticate_user(data['password']):
		return {'message': 'Login successful'}, 200
	else:
		return {'message': 'Invalid username or password'}, 401

@app.route('/view_profile', methods=['GET'])
def view_profile():
	data = request.get_json()
	profile = Profile(data['username'])
	return profile.view_profile(), 200

@app.route('/edit_profile', methods=['POST'])
def edit_profile():
	data = request.get_json()
	profile = Profile(data['username'])
	profile.edit_profile(data['new_info'])
	return {'message': 'Profile edited successfully'}, 200

@app.route('/list_books', methods=['GET'])
def list_books():
	data = request.get_json()
	profile = Profile(data['username'])
	return {'books': profile.list_books()}, 200

@app.route('/follow_user', methods=['POST'])
def follow_user():
	data = request.get_json()
	profile = Profile(data['username'])
	profile.follow_user(data['user_to_follow'])
	return {'message': 'User followed successfully'}, 200

@app.route('/create_book_club', methods=['POST'])
def create_book_club():
	data = request.get_json()
	book_club = BookClub(data['name'], data['admin'])
	book_club.create_book_club()
	return {'message': 'Book club created successfully'}, 201

@app.route('/join_book_club', methods=['POST'])
def join_book_club():
	data = request.get_json()
	book_club = BookClub(data['name'], data['admin'])
	book_club.add_member(data['member'])
	return {'message': 'Member added successfully'}, 200

@app.route('/schedule_meeting', methods=['POST'])
def schedule_meeting():
	data = request.get_json()
	meeting = Meeting(data['meeting_id'], data['book_club'], data['date'], data['time'])
	meeting.schedule_meeting()
	return {'message': 'Meeting scheduled successfully'}, 201

@app.route('/create_discussion', methods=['POST'])
def create_discussion():
	data = request.get_json()
	discussion = Discussion(data['discussion_id'], data['book_club'], data['topic'])
	discussion.create_discussion()
	return {'message': 'Discussion created successfully'}, 201

@app.route('/post_comment', methods=['POST'])
def post_comment():
	data = request.get_json()
	discussion = Discussion(data['discussion_id'], data['book_club'], data['topic'])
	discussion.post_comment(data['comment'])
	return {'message': 'Comment posted successfully'}, 200

@app.route('/upload_resource', methods=['POST'])
def upload_resource():
	data = request.get_json()
	discussion = Discussion(data['discussion_id'], data['book_club'], data['topic'])
	discussion.upload_resource(data['resource'])
	return {'message': 'Resource uploaded successfully'}, 200

@app.route('/suggest_book', methods=['POST'])
def suggest_book():
	data = request.get_json()
	book = Book(data['book_id'], data['title'], data['author'])
	book.suggest_book()
	profile = Profile(data['username'])
	profile.books.append(data['book_id'])
	return {'message': 'Book suggested successfully'}, 201

@app.route('/vote_for_book', methods=['POST'])
def vote_for_book():
	data = request.get_json()
	book = Book(data['book_id'], data['title'], data['author'])
	book.vote_for_book()
	return {'message': 'Vote cast successfully'}, 200

@app.route('/generate_recommendation', methods=['GET'])
def generate_recommendation():
	data = request.get_json()
	recommendation = Recommendation(data['username'])
	return {'recommendation': recommendation.generate_recommendation()}, 200

@app.route('/highlight_popular_book', methods=['GET'])
def highlight_popular_book():
	recommendation = Recommendation('')
	return {'popular_book': recommendation.highlight_popular_book()}, 200

@app.route('/admin/manage_book_club', methods=['POST'])
def admin_manage_book_club():
	data = request.get_json()
	admin = Admin(data['username'])
	return {'message': admin.manage_book_club(data['book_club_name'], data['action'], data.get('member'))}, 200

@app.route('/admin/manage_user_account', methods=['POST'])
def admin_manage_user_account():
	data = request.get_json()
	admin = Admin(data['username'])
	return {'message': admin.manage_user_account(data['user_to_manage'], data['action'])}, 200

@app.route('/admin/remove_inappropriate_content', methods=['POST'])
def admin_remove_inappropriate_content():
	data = request.get_json()
	admin = Admin(data['username'])
	return {'message': admin.remove_inappropriate_content(data['discussion_id'])}, 200

@app.route('/admin/view_analytics', methods=['GET'])
def admin_view_analytics():
	data = request.get_json()
	admin = Admin(data['username'])
	return admin.view_analytics(), 200

@app.route('/setup_notification', methods=['POST'])
def setup_notification():
	data = request.get_json()
	notification = Notification(data['username'], data['message'])
	notification.setup_notification()
	return {'message': 'Notification set up successfully'}, 200

@app.route('/send_email_alert', methods=['POST'])
def send_email_alert():
	data = request.get_json()
	notification = Notification(data['username'], data['message'])
	notification.send_email_alert()
	return {'message': 'Email alert sent'}, 200

if __name__ == '__main__':
	app.run(debug=True)
